+++ packages/world/src/systems/SpawnSystem.sol
@@ -22,7 +22,7 @@ import { checkWorldStatus, getUniqueEntity, gravityApplies, inWorldBorder } from
 import { transferAllInventoryEntities } from "../utils/InventoryUtils.sol";
 import { notify, SpawnNotifData } from "../utils/NotifUtils.sol";
 import { getForceField } from "../utils/ForceFieldUtils.sol";
-
+import { TerrainLib } from "./libraries/TerrainLib.sol";
 import { EntityId } from "../EntityId.sol";
 
 contract SpawnSystem is System {
@@ -50,13 +50,17 @@ contract SpawnSystem is System {
 
     EntityId playerEntityId = getUniqueEntity();
     EntityId existingEntityId = ReversePosition._get(spawnCoord.x, spawnCoord.y, spawnCoord.z);
-    require(existingEntityId.exists(), "Cannot spawn on an unrevealed block");
-    require(ObjectType._get(existingEntityId) == AirObjectID, "Cannot spawn on a non-air block");
+    if (!existingEntityId.exists()) {
+      uint16 terrainObjectTypeId = TerrainLib._getBlockType(spawnCoord);
+      require(terrainObjectTypeId == AirObjectID, "Cannot spawn on a non-air block");
+    } else {
+      require(ObjectType._get(existingEntityId) == AirObjectID, "Cannot spawn on a non-air block");
 
-    // Transfer any dropped items
-    transferAllInventoryEntities(existingEntityId, playerEntityId, PlayerObjectID);
+      // Transfer any dropped items
+      transferAllInventoryEntities(existingEntityId, playerEntityId, PlayerObjectID);
 
-    Position._deleteRecord(existingEntityId);
+      Position._deleteRecord(existingEntityId);
+    }
 
     // Create new entity
     Position._set(playerEntityId, spawnCoord.x, spawnCoord.y, spawnCoord.z);
